+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Fixes
 
 - Leave `inApp` flag for stack frames undecided in SDK if unsure and let ingestion decide instead ([#2547](https://github.com/getsentry/sentry-java/pull/2547))
+- Allow `0.0` error sample rate ([#2573](https://github.com/getsentry/sentry-java/pull/2573))
 
 ## 6.14.0